Update extension from js to ts
