+++ advancewars/typings.d.ts
@@ -1,0 +1,7 @@
+declare module '*.css';
+
+interface COSaveData {
+    alias: string;
+    customUrl: string;
+    lostUrl: string;
+}
